Only test appropriate CIDR ranges
If we're examining an IPv4 address it doesn't make sense to
test whether that is included in our IPv6-specific blacklists.

So we parse our known-bad/known-local ranges into a pair of
maps, one for IPv4 and one for IPv6.  We use the appropriate one
when we're doing the lookup, to ensure we avoid wasting time
doing tests that will never succeed.

This closes #6.